Avoid confusing IDs in error summary examples
We’ve had feedback that the IDs used in these examples might suggest that the links should go to the error message rather than the input.

To make this clearer, avoid suffixing the link targets with the word ‘error’, and use `input` where it’d be realistic a user might do so, like for simple text inputs.

In the case of date inputs, the date input component uses the field name as a suffix, so it’s likely the field ID will end with e.g. ‘-day’, so don’t add a further suffix there.

For the checkbox example, use an ID which makes it clear that it links to the first checkbox.